# docs/swagger.yaml
definitions:
...
  hei-gin_plugins_plugin-client_session.SessionExitParam:
    properties:
      user_id:
        type: string
    type: object
  hei-gin_plugins_plugin-client_session.SessionExitTokenParam:
    properties:
      token:
        type: string
      user_id:
        type: string
    type: object
  hei-gin_plugins_plugin-client_user.UpdateAvatarParam:
    properties:
      avatar:
        type: string
    type: object
  hei-gin_plugins_plugin-client_user.UpdatePasswordParam:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    type: object
  hei-gin_plugins_plugin-client_user.UpdateProfileParam:
    properties:
      avatar:
        type: string
      email:
        type: string
      nickname:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
...
  user.ClientUserVO:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      gender:
        type: string
      github:
        type: string
      id:
        type: string
      last_login_at:
        type: string
      last_login_ip:
        type: string
      login_count:
        type: integer
      motto:
        type: string
      nickname:
        type: string
      password:
        type: string
      phone:
        type: string
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  utils.IdsParam:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
info:
  contact:
    name: Charlie
    url: https://github.com/jiangbyte
  description: Hei Gin 后台管理系统 API 文档
  license:
    name: MIT
    url: https://github.com/jiangbyte/hei-gin/blob/main/LICENSE
  termsOfService: https://github.com/jiangbyte/hei-gin
  title: Hei Gin API
  version: 1.0.0
paths:
...
  /api/v1/c/client-user/update-avatar:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/c/client-user/update-avatar，C端用户更新头像
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/hei-gin_plugins_plugin-client_user.UpdateAvatarParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: C端用户更新头像
      tags:
      - C端用户
  /api/v1/c/client-user/update-password:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/c/client-user/update-password，C端用户修改密码
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/hei-gin_plugins_plugin-client_user.UpdatePasswordParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: C端用户修改密码
      tags:
      - C端用户
  /api/v1/c/client-user/update-profile:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/c/client-user/update-profile，C端用户更新个人信息
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/hei-gin_plugins_plugin-client_user.UpdateProfileParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: C端用户更新个人信息
      tags:
      - C端用户
...
  /api/v1/client-user/create:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/client-user/create，C端用户创建
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.ClientUserVO'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: C端用户创建
      tags:
      - C端用户
  /api/v1/client-user/detail:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/client-user/detail，C端用户详情查询
      parameters:
      - description: id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: C端用户详情查询
      tags:
      - C端用户
  /api/v1/client-user/modify:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/client-user/modify，C端用户修改
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.ClientUserVO'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: C端用户修改
      tags:
      - C端用户
  /api/v1/client-user/page:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/client-user/page，C端用户分页查询
      parameters:
      - in: query
        name: current
        type: integer
      - in: query
        name: keyword
        type: string
      - in: query
        name: size
        type: integer
      - in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: C端用户分页查询
      tags:
      - C端用户
  /api/v1/client-user/remove:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/client-user/remove，C端用户删除
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/utils.IdsParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: C端用户删除
      tags:
      - C端用户
...
  /api/v1/client/session/exit:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/client/session/exit，C端会话强退会话
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/hei-gin_plugins_plugin-client_session.SessionExitParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: C端会话强退会话
      tags:
      - C端会话
  /api/v1/client/session/exit-token:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/client/session/exit-token，C端会话强退令牌
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/hei-gin_plugins_plugin-client_session.SessionExitTokenParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: C端会话强退令牌
      tags:
      - C端会话
  /api/v1/client/session/page:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/client/session/page，C端会话分页查询
      parameters:
      - in: query
        name: current
        type: integer
      - in: query
        name: keyword
        type: string
      - in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: C端会话分页查询
      tags:
      - C端会话
  /api/v1/client/session/tokens:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/client/session/tokens，C端会话令牌列表
      parameters:
      - description: user_id
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: C端会话令牌列表
      tags:
      - C端会话
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"